refactor(core): extract stereo channel transcription
Move transcribe_channels, transcribe_channels_inner, and
transcribe_channel_streams into engine/channels.rs. Public
transcribe_* wrappers stay in transcribe.rs.
